+++ 006/ThanhNguyen/Exercise6/Exercise6/Screen/Login/CustomView/CheckBoxView.swift
@@ -1,0 +1,20 @@
+//
+//  CheckBoxView.swift
+//  Exercise6
+//
+//  Created by Thanh Nguyen X. [4] VN.Danang on 29/08/2023.
+//
+
+import SwiftUI
+
+struct CheckBoxView: View {
+    @Binding var checked: Bool
+
+    var body: some View {
+        Image(systemName: checked ? "checkmark.square.fill" : "square")
+            .foregroundColor( checked ? .blue : .secondary)
+            .onTapGesture {
+                self.checked.toggle()
+            }
+    }
+}
