+++ 006/ThanhNguyen/Exercise6/Exercise6/Screen/Login/LoginViewModel.swift
@@ -9,12 +9,6 @@ import Foundation
 
 final class LoginViewModel: ObservableObject {
 
-    private struct Constants {
-        // Minimum 8 characters at least 1 Uppercase Alphabet, 1 Lowercase Alphabet, 1 Number and 1 Special Character
-        static let passwordRegex = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[d$@$!%*?&#])[A-Za-z\\dd$@$!%*?&#]{8,}"
-        static let emailRegex = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,64}"
-    }
-
     @Published var email: String = ""
     @Published var password: String = ""
     @Published var errorTitle: String = ""
@@ -46,13 +40,5 @@ final class LoginViewModel: ObservableObject {
     func validateSignInButton() -> Bool {
         return email.isMatch(regex: Constants.emailRegex)
         && password.isMatch(regex: Constants.passwordRegex)
-    }
-}
-
-extension String {
-    public func isMatch(regex: String) -> Bool {
-        guard let regex = try? NSRegularExpression(pattern: regex) else { return false }
-        let range = NSRange(location: 0, length: self.utf16.count)
-        return regex.firstMatch(in: self, range: range) != nil
     }
 }
