+++ 006/ThanhNguyen/Exercise6/Exercise6/Screen/Welcome/WelcomeViewModel.swift
@@ -7,43 +7,43 @@
 
 import Foundation
 
-final class WelcomeViewModel: ObservableObject {
+struct WelcomeData {
+    let title: String
+    let subtitle: String
+    let imageName: String
+}
 
-    struct Constants {
-        static let defaultPage: Int = 0
-    }
-
-    struct WelcomeData {
-        let title: String
-        let subtitle: String
-        let imageName: String
-    }
-
-    let welcomePages: [WelcomeData] = [
+extension WelcomeData {
+    static let welcomePages: [WelcomeData] = [
         .init(title: "Explore the world easily", subtitle: "To your desire", imageName: "img_welcome1"),
         .init(title: "Reach the unknown spot", subtitle: "To your destination", imageName: "img_welcome2"),
         .init(title: "Make connects with Travello", subtitle: "To your dream trip", imageName: "img_welcome3")
     ]
+}
 
-    @Published var currentPage: Int
+final class WelcomeViewModel: ObservableObject {
+
+    @Published var currentPage: Int {
+        didSet {
+            welcomeData = WelcomeData.welcomePages[currentPage]
+        }
+    }
     @Published var welcomeData: WelcomeData
 
     init() {
         self.currentPage = Constants.defaultPage
-        self.welcomeData = welcomePages[Constants.defaultPage]
+        self.welcomeData = WelcomeData.welcomePages[Constants.defaultPage]
     }
 
     func handleTapBackPage() {
         if currentPage > Constants.defaultPage {
             currentPage -= 1
-            welcomeData = welcomePages[currentPage]
         }
     }
 
     func handleTapNextPage() -> Bool {
-        if currentPage < welcomePages.count - 1 {
+        if currentPage < WelcomeData.welcomePages.count - 1 {
             currentPage += 1
-            welcomeData = welcomePages[currentPage]
             return false
         } else {
             return true
